replace caper_uri with autouri, update doc
