+++ bin/qc2tsv
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-
 try:
     import qc2tsv
 except:
@@ -7,8 +6,8 @@ except:
     script_path = os.path.dirname(os.path.realpath(__file__))
     sys.path.append(os.path.join(script_path, '../'))
     import qc2tsv
+from qc2tsv.cli import main
 
-from qc2tsv.qc2tsv import main
 
-main()
-
+if __name__ == '__main__':
+    main()
